pyproject.toml
[project]
name = "databricks-pypi-installer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk>=0.36.0",
    "pip>=24.3.1",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "pytest>=8.3.3",
]
